break out a file
